requirements.txt
amqp==2.5.1
argon2-cffi==19.1.0
asn1crypto==0.24.0
astroid==2.2.5
celery==4.3.0
cffi==1.12.3
chardet==3.0.4
CommonMark==0.7.4
cryptography==2.7
dateparser==0.7.2
defusedxml==0.6.0
Deprecated==1.2.6
Django==1.11.24
django-admin-sortable2==0.7.3
django-extensions==2.2.1
django-historylinks==1.1.1
django-jet==1.0.8
django-jinja==2.4.1
django-reversion==2.0.9
django-sortedm2m==1.5.0
django-watson==1.5.2
django-webpack-loader==0.6.0
docopt==0.6.2
ecdsa==0.13.2
Fabric3==1.14.post1
fabric3-virtualenv==0.3.1
future==0.17.1
html2text==2018.1.9
idna==2.8
importlib-metadata==0.23
isort==4.3.20
Jinja2==2.10.1
kombu==4.6.4
lazy-object-proxy==1.4.2
lxml==4.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
oauthlib==3.1.0
onespacemedia-cms==4.0.7
onespacemedia-server-management==3.2.2
osm-jet==1.0.6
paramiko==1.17.6
Pillow==6.1.0
psycopg2==2.8.3
pycparser==2.19
pycrypto==2.6.1
PyGithub==1.43.8
PyJWT==1.7.1
pylint==2.3.0
pylint-django==2.0.5
pylint-plugin-utils==0.5
python-dateutil==2.8.0
python-magic==0.4.15
python-memcached==1.59
python3-openid==3.1.0
pytz==2019.2
redis==3.3.8
regex==2019.8.19
requests==2.22.0
requests-oauthlib==1.2.0
rollbar==0.14.7
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sorl-thumbnail==12.5.0
tinypng==3.0.0
typed-ast==1.4.0
tzlocal==2.0.0
urllib3==1.25.3
vine==1.3.0
Werkzeug==0.15.6
wrapt==1.11.2
zipp==0.6.0
